requirements.txt
Django
djangorestframework
flake8
django-cors-headers
djangorestframework-simplejwt